Converted authenticator class to abstract class
